requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
pydantic>=2.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
fastapi>=0.115.0
uvicorn>=0.34.0
faiss-cpu>=1.9.0
httpx>=0.28.0
python-multipart>=0.0.18
gunicorn>=21.2.0
